// my-app/src/types.ts
export default interface Option {
    id: number;
    name: string;
  }
  
  export type cardInfoProps = {
    id: number;
    price: number;
    weight: number;
    city_production: string;
    category: string;
    image_path: string;
    manufacturing_ccompany: string;
    status: number;
    component_name: string;
    description: string;
  };